@@ -5,23 +5,20 @@
     "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
-        "guzzlehttp/guzzle": "^7.8",
-        "laravel/framework": "^10.48",
-        "laravel/sanctum": "^3.3",
+        "php": "^8.2",
+        "laravel/framework": "^11.13",
         "laravel/tinker": "^2.9",
         "livewire/livewire": "^3.5",
         "wire-elements/modal": "^2.0"
     },
     "require-dev": {
-        "fakerphp/faker": "^1.23.1",
-        "laravel/breeze": "^1.28",
+        "fakerphp/faker": "^1.23",
+        "laravel/breeze": "^2.1",
         "laravel/pint": "^1.16",
         "laravel/sail": "^1.30",
-        "mockery/mockery": "^1.6.12",
-        "nunomaduro/collision": "^7.10",
-        "phpunit/phpunit": "^10.5",
-        "spatie/laravel-ignition": "^2.8"
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.1",
+        "phpunit/phpunit": "^11.2"
     },
     "autoload": {
         "psr-4": {
@@ -47,7 +44,9 @@
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "@php artisan key:generate --ansi"
+            "@php artisan key:generate --ansi",
+            "@php -r \"file_exists('database/database.sqlite') || touch('database/database.sqlite');\"",
+            "@php artisan migrate --graceful --ansi"
         ]
     },
     "extra": {
