@@ -9,7 +9,7 @@ return [
     |
     | The following language lines are the default lines which match reasons
     | that are given by the password broker for a password update attempt
-    | has failed, such as for an invalid token or invalid new password.
+    | outcome such as failure due to an invalid password / reset token.
     |
     */
 
